pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "text2sql"
version = "0.1.0"
description = "LangChain agents and utilities for DeepSeek-powered text-to-SQL experiments."
readme = "readme.txt"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "loguru>=0.7.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
